--- src/context/modeProvider.jsx
@@ -1,11 +1,0 @@
-import { createContext, useContext } from "react";
-
-export const modeContext = createContext();
-export const useMode = () => {
-  return useContext(modeContext);
-};
-
-export const setModeContext = createContext();
-export const useSetMode = () => {
-  return useContext(setModeContext);
-};
